requirements.txt
pandas>=2.1.3
numpy>=1.24.0
langchain>=0.1.16
langchain-community>=0.0.37
sentence-transformers>=2.2.0
langchain-text-splitters==0.0.1
sentence-transformers==2.2.2
chromadb>=0.4.24
transformers>=4.30.0
torch>=2.1.2
google-genai==0.3.0
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
scikit-learn>=1.3.0
beautifulsoup4==4.12.2c
python-dotenv==1.0.0
PyPDF2==3.0.0
